Title: 실수 입력받아 계산 2, Time: 6ms, Memory: 0MB, Status: Passed - CodeTree

+++ 240506/실수 입력받아 계산 2/calculation-by-inputting-a-real-number-2.c
@@ -1,8 +1,8 @@
 #include <stdio.h>
 
 int main() {
-    float a;
-    scanf("%f", &a);
-    printf("%.2f", a+1.5);
+    double a;
+    scanf("%lf", &a);
+    printf("%.2lf", a+1.5);
     return 0;
 }
